+++ CHANGELOG.md
@@ -6,5 +6,6 @@
 * Enabled travis.
 * Added error? to check if the command ran successfully. MIXLIB-18.
 * Remove GC.disable hack for non-ruby 1.8.8
+* Handle ESRCH from getpgid of a zombie on OS X
 
 ## Last Release: 1.3.0 (12/03/2013)